refactor: add unit tests for the layout utility
